+++ src/lib/widgets/index.ts
@@ -1,1 +1,2 @@
 export * from './custom-widget/custom-widget.directive';
+export * from './layers-toggle/layers-toggle.component';